requirements.txt
alembic==1.16.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email-validator==1.3.1
Flask==2.3.2
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==2.3.7
WTForms==3.2.1
python-dotenv==1.0.0
Pillow==10.0.1
authlib==1.6.0
requests==2.32.4
Flask-Limiter>=3.4.0

